refactor(discover): an expand control, not a row in the kind menu
Every kind in that menu is a corpus you already own. Listing Discover among them
said it was one more of your things, which is the one thing it is not — and it
needed a sentence of its own ("What others have shared") to say otherwise.

An expand says it without words: the same surface, wider. It wears the exact
counterpart of the control that collapses the panel, sits at the other end of
the same header, and leaves the kind menu to name kinds.

The panel still closes behind it — two browse surfaces stacked is one too many.
